requirements/runtime_cuda.txt
accelerate>=0.29.3
aiohttp
cloudpickle
einops
fastapi
fire
mmengine-lite
numpy
openai
openai_harmony
partial_json_parser
peft<=0.14.0
pillow
prometheus_client
protobuf
pybase64
pydantic>2.0.0
pyzmq
ray
safetensors
sentencepiece
shortuuid
tiktoken
torch<=2.8.0,>=2.0.0
torchvision<=0.23.0,>=0.15.0
transformers==4.57.1
triton<=3.4.0,>=3.0.0; sys_platform == "linux" and "aarch64" not in platform_machine and "arm" not in platform_machine
uvicorn
xgrammar
uvloop
datasets>=4.4.1